pyproject.toml
[project]
name = "datagouv-mcp"
dynamic = ["version"]
description = "Model Context Protocol (MCP) server for interacting with data.gouv.fr datasets and resources via LLM chatbots"
authors = [
    {name = "data.gouv.fr", email = "opendatateam@data.gouv.fr"}
]
readme = "README.md"
requires-python = ">=3.13,<3.15"
dependencies = [
    "httpx>=0.28.0",
    "mcp>=1.25.0,<2",
    "pyyaml>=6.0",
]

[build-system]
requires = ["setuptools>=61.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.14.0",
    "ty>=0.0.18",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests that call live remote services",
]
